refactor(atomic): move components to their own files and reusable stuff to common file
https://coveord.atlassian.net/browse/KIT-1517
